+++ src/lib/constants.ts
@@ -1,1 +1,3 @@
 export const middleOfUSA = [-100, 40];
+export const routeMin = 1;
+export const routeMax = 10;
